Correção da declaração int para double

+++ cap3/Casting.java
@@ -5,9 +5,10 @@ class Casting {
     double d = 3.1415;
     double i = d;
 
-    // tambem nao compila
+    // tambem nao compila pois fica dificil para o compilador
+    // determinar se o valor estara dentro desse double em tempo de execucao
     double a = 5;
-    int b = a;
+    double b = a;
 
     // mas assim compila
     int c = 5;
